Scaffolded device infrastructure
